2020_10_23 (重大)フレームワーク最適化(モデル、画像)

+++ Particle.cpp
@@ -28,7 +28,7 @@ void ParticleSystem::Create(ParitcleSetting* setting) {
 
 	// パーティクル資料生成
 	mParticleAmount = setting->Amount;
-	mParticleLifeMax = setting->LifeMinMax.y;
+	mParticleLifeMax = (int)setting->LifeMinMax.y;
 	mparticle = new Particle[setting->Amount];
 	mVel = new D3DXVECTOR3[setting->Amount];
 	mlife = new float[setting->Amount];
